requirements.txt
mss==10.1.0
pillow==12.1.1
watchdog==6.0.0
flask
requests
pytesseract
Pillow

